move durable agent into ee
