+++ packages/tracecat-ee/tracecat_ee/agent/__init__.py
@@ -1,0 +1,1 @@
+"""Tracecat Enterprise Edition agent features."""
