+++ packages/tracecat-ee/tracecat_ee/agent/activities.py
@@ -17,7 +17,6 @@ from pydantic_ai.messages import (
 from pydantic_ai.models import ModelRequestParameters
 from temporalio import activity
 
-from tracecat.agent.context import AgentContext
 from tracecat.agent.models import (
     ModelInfo,
     ModelRequestArgs,
@@ -41,6 +40,7 @@ from tracecat.dsl.enums import PlatformAction
 from tracecat.logger import logger
 from tracecat.secrets import secrets_manager
 from tracecat.types.auth import Role
+from tracecat_ee.agent.context import AgentContext
 
 
 class InvokeToolArgs(BaseModel):
