+++ packages/tracecat-ee/tracecat_ee/agent/approvals/__init__.py
@@ -1,0 +1,1 @@
+"""Agent approval system."""
